remove hardcoding of true flag for "keycloak.enabled"
If Keycloak dependency is added : now it is must to give "keycloak.enabled=false", If we do not wish to use keycloak however still the project has dependency added.
The usecase comes when we want to create a project jar having many IDPs switching on the basis of key's value. 
If we have matchIfMissing =false instead of true, only the one using keycloak can add keycloack.enabled=true.